consolidate shared types
