+++ typescript/mcp/src/resources/weeklyMealPlan.ts
@@ -1,15 +1,7 @@
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { McpError } from '@modelcontextprotocol/sdk/types.js';
 import { API } from '../utils.js';
-
-export interface WeeklyMealPlan {
-  days: {
-    date: string;
-    mealId: number;
-    mealName: string;
-    effort: 'LOW' | 'MED' | 'HIGH';
-  }[];
-}
+import type { WeeklyMealPlan } from '@meal-planner/shared/dist/types';
 
 export async function fetchWeeklyMealPlan(): Promise<WeeklyMealPlan> {
   const response = await fetch(`${API}/api/mealplan`);
