+++ typescript/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './days';
+export * from './types';
