+++ typescript/shared/package.json
@@ -1,8 +1,8 @@
 {
   "name": "@meal-planner/shared",
   "version": "1.0.0",
-  "main": "dist/days.js",
-  "types": "dist/days.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "private": true,
   "scripts": {
     "build": "tsc"
