--- typescript/ui/src/constants/days.ts
@@ -1,10 +1,0 @@
-export const DAYS_OF_THE_WEEK = [
-  'Monday',
-  'Tuesday',
-  'Wednesday',
-  'Thursday',
-  'Friday',
-  'Saturday',
-  'Sunday',
-] as const;
-export type DayOfTheWeek = (typeof DAYS_OF_THE_WEEK)[number];
